backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pyyaml==6.0.1

# ScyllaDB Driver
scylla-driver==3.29.3

# WebSocket Support
websockets==12.0
python-multipart==0.0.6

# Embeddings & LLM
requests==2.31.0
httpx==0.26.0

# Text Processing
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2
pymupdf==1.23.8

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
numpy==1.26.3
uuid==1.30

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
